+++ dedup_comp/host/Enclave_u.c
@@ -14,8 +14,8 @@ typedef struct ms_ocall_load_text_file_t {
 } ms_ocall_load_text_file_t;
 
 typedef struct ms_ocall_get_time_t {
-	long int* ms_second;
-	long int* ms_nanosecond;
+	int* ms_second;
+	int* ms_nanosecond;
 } ms_ocall_get_time_t;
 
 typedef struct ms_ocall_request_find_t {
